Add 'readable' Hawaiian shirts
Functionally similar to reading a t-shirt or apron, but rather than
actual text printed on the shirt being displayed, the design of the
Hawaiian shirt is described: for example, "hula dancers on an orange
background" or "tropical fish on an abstract background".  Much like
t-shirts have their text included in the game-end inventory list ('a
blessed +2 t-shirt with text "foo"'), Hawaiian shirts now have a brief
description of their design appended to the item name when the hero's
possessions are identified after the end of the game.

Because 'reading' a Hawaiian shirt doesn't actually involve reading
text, using the 'r' commnad in this way doesn't break illiterate
conduct.